refactor: rename moon-themed surface rule constants to irradiated naming
Replace leftover MOON_*/MARE/LUNASLATE identifiers with domain-appropriate
names (IS_IRRADIATED_PLAIN, ENRICHED_SOUL_SOIL, LEAD_ROCK, etc.) and update
the noise settings random_name to match. Also changes heat accumulator to
double, uses getF() for the cooler proxy config, switches the alarm to
REACTOR_ALARM_2, adds RodType.toString(), and removes CNSounds.java.

+++ src/main/java/net/nuclearteam/createnuclear/content/multiblock/reactorLogic/DefaultHeatCalculator.java
@@ -28,7 +28,7 @@ public class DefaultHeatCalculator implements IHeatCalculator {
 
     @Override
     public double computeHeat(BigItemStack bigFuelItem, BigItemStack bigCoolerItem, BigFluidStack bigFluidStack, ReactorFluidType type, int countGraphiteRod, int countUraniumRod, ReactorControllerInventory inventory, double overHeat) {
-        int heat = 0;
+        double heat = 0;
 
         ListTag list = inventory.getStackInSlot(0).getOrCreateTag().getCompound("pattern").getList("Items", Tag.TAG_COMPOUND);
